Repackaged project contents.
