+++ ProjectThemis/src/client/Connect4Client.java
@@ -1,3 +1,4 @@
+package client;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Font;
