+++ ProjectThemis/src/client/Menu.java
@@ -1,3 +1,4 @@
+package client;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Component;
