+++ ProjectThemis/src/client/ProjectThemisClient.java
@@ -1,5 +1,8 @@
+package client;
 import java.io.*;
 import java.net.*;
+
+import shared.PasswordAuthentication;
 
 //Main class for client side things.
 //Will create client classes for each game.
