+++ ProjectThemis/src/client/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package contains all classes that 
+ */
+/**
+ * @author Devin
+ *
+ */
+package client;
