+++ ProjectThemis/src/server/MySQLWrapper.java
@@ -1,3 +1,4 @@
+package server;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
