+++ ProjectThemis/src/server/ProjectThemisServer.java
@@ -1,3 +1,4 @@
+package server;
 
 import java.net.*;
 import java.io.*;
