+++ ProjectThemis/src/server/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Contains all the code that should be sent with ProjectThemisServer.java
+ */
+/**
+ * @author Devin
+ *
+ */
+package server;
