+++ ProjectThemis/src/shared/PasswordAuthentication.java
@@ -1,3 +1,4 @@
+package shared;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.spec.InvalidKeySpecException;
