+++ ProjectThemis/src/shared/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Contains all the code that should be sent with ProjectThemisClient.java
+ */
+/**
+ * @author Devin
+ *
+ */
+package shared;
